menu framework update 1.2
added the conditions for various gamemodes
successfully added the createMenu function. However it only creates the button without text.
drawOver() func for moving objects. may try to put it in the func file in the future.
fixed circular import
additional minor changes.
still need to check for where the user clicks in order to navigate menu